+++ R/Driver.R
@@ -1,12 +1,13 @@
 #' @useDynLib RSQLite
 #' @importFrom Rcpp sourceCpp
-#' @include Connection.R
 NULL
 
 #' Class SQLiteDriver (and methods).
 #' 
 #' The SQLiteDriver, which is used to select the correct method in 
-#' \code{dbConnect()}. See more details in \code{\link{SQLite}}
+#' \code{dbConnect()}. See more details in \code{\link{SQLite}}.
+#' It is used purely for dispatch and \code{dbUnloadDriver} is unnecessary
+#' (and a null-op).
 #' 
 #' @keywords internal
 #' @export
@@ -15,17 +16,18 @@ setClass("SQLiteDriver",
 )
 
 #' @rdname SQLiteDriver-class
+#' @export
 setMethod("show", "SQLiteDriver", function(object) {
   cat("<SQLiteDriver>\n")
 })
 
-#' Unload SQLite driver.
-#' 
-#' This function is no longer needed.
-#' 
-#' @param drv Object created by \code{\link{SQLite}}
-#' @param ... Ignored. Needed for compatibility with generic.
-#' @keywords internal
+#' @rdname SQLiteDriver-class
+#' @export
+setMethod("dbIsValid", "SQLiteDriver", function(dbObj) {
+  TRUE
+})
+
+#' @rdname SQLiteDriver-class
 #' @export
 setMethod("dbUnloadDriver", "SQLiteDriver", function(drv, ...) {
   invisible(TRUE)